possibly wrong url behind "Next!"
The "Next!" link leads to the article "Extra - Navigation". The tutorial there expects an existing implementation for an android project, which is not the case when you just followed along the "Core"-tutorial. Additionally, the navigation-tutorial is the last in the list, while android is the next. So I assume the next tutorial should be the android tutorial? At least this navigation thing got me very confused.
Cheers!